requirements.txt
streamlit
pandas
gspread_pandas
google-api-python-client
json
os